freebsd

+++ ddl/ingest/rlimit_other.go
@@ -29,7 +29,7 @@ func genRLimit() uint64 {
 	if err != nil {
 		logutil.BgLogger().Warn(LitErrGetSysLimitErr, zap.Error(err), zap.String("default", "1024"))
 	} else {
-		rLimit = rl.Cur
+		rLimit = uint64(rl.Cur)
 	}
 	return rLimit
 }
